requirements.txt
annotated-types==0.7.0
anyio==4.6.2
async-timeout==4.0.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
joblib==1.4.2
numpy==1.24.4
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic-core==2.23.4
pymongo==4.10.1
# Pin pymilvus + grpcio: unpinned pymilvus>=2.3 pulled 2.6.x + grpcio 1.80 which could hang on Windows at MilvusClient connect.
pymilvus==2.4.9
grpcio==1.65.5
# pymilvus 2.4.x -> environs -> marshmallow; marshmallow 4.x breaks environs.
marshmallow>=3.21,<4
# pymilvus 2.4.x imports pkg_resources; setuptools 82+ may omit it.
setuptools>=69.0.0,<81
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
threadpoolctl==3.5.0
tomli==2.0.2
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
webencodings==0.5.1
